# sdk/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DND Booker API",
    "version": "1.0.0-alpha"
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/agent-runs/{runId}/checkpoints": {
      "get": {
        "operationId": "listAgentCheckpoints",
        "summary": "List agent checkpoints.",
        "tags": [
          "agentRuns"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "runId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "label": {
                        "type": "string"
                      },
                      "summary": {
                        "type": "string",
                        "nullable": true
                      },
                      "cycleIndex": {
                        "type": "integer"
                      },
                      "isBest": {
                        "type": "boolean"
                      },
                      "scorecard": {
                        "type": "object",
                        "properties": {
                          "overallScore": {
                            "type": "number"
                          },
                          "exportScore": {
                            "type": "number",
                            "nullable": true
                          },
                          "blockingFindingCount": {
                            "type": "integer"
                          },
                          "warningFindingCount": {
                            "type": "integer"
                          },
                          "utilityDensityAverage": {
                            "type": "number",
                            "nullable": true
                          },
                          "sparsePageCount": {
                            "type": "integer"
                          },
                          "thinRandomTableCount": {
                            "type": "integer"
                          },
                          "lowUtilityDensityCount": {
                            "type": "integer"
                          },
                          "suspiciousStatBlockCount": {
                            "type": "integer"
                          },
                          "generatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "summary": {
                            "type": "string"
                          },
                          "latestExportJobId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          }
                        },
                        "required": [
                          "overallScore",
                          "exportScore",
                          "blockingFindingCount",
                          "warningFindingCount",
                          "utilityDensityAverage",
                          "sparsePageCount",
                          "thinRandomTableCount",
                          "lowUtilityDensityCount",
                          "suspiciousStatBlockCount",
                          "generatedAt",
                          "summary",
                          "latestExportJobId"
                        ],
                        "additionalProperties": false,
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "runId",
                      "label",
                      "summary",
                      "cycleIndex",
                      "isBest",
                      "scorecard",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/projects/{projectId}/agent-runs/{runId}/checkpoints/{checkpointId}/restore": {
      "post": {
        "operationId": "restoreAgentCheckpoint",
        "summary": "Restore an agent checkpoint.",
        "tags": [
          "agentRuns"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "runId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "label": {
                      "type": "string"
                    },
                    "summary": {
                      "type": "string",
                      "nullable": true
                    },
                    "cycleIndex": {
                      "type": "integer"
                    },
                    "isBest": {
                      "type": "boolean"
                    },
                    "scorecard": {
                      "type": "object",
                      "properties": {
                        "overallScore": {
                          "type": "number"
                        },
                        "exportScore": {
                          "type": "number",
                          "nullable": true
                        },
                        "blockingFindingCount": {
                          "type": "integer"
                        },
                        "warningFindingCount": {
                          "type": "integer"
                        },
                        "utilityDensityAverage": {
                          "type": "number",
                          "nullable": true
                        },
                        "sparsePageCount": {
                          "type": "integer"
                        },
                        "thinRandomTableCount": {
                          "type": "integer"
                        },
                        "lowUtilityDensityCount": {
                          "type": "integer"
                        },
                        "suspiciousStatBlockCount": {
                          "type": "integer"
                        },
                        "generatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "summary": {
                          "type": "string"
                        },
                        "latestExportJobId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        }
                      },
                      "required": [
                        "overallScore",
                        "exportScore",
                        "blockingFindingCount",
                        "warningFindingCount",
                        "utilityDensityAverage",
                        "sparsePageCount",
                        "thinRandomTableCount",
                        "lowUtilityDensityCount",
                        "suspiciousStatBlockCount",
                        "generatedAt",
                        "summary",
                        "latestExportJobId"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "runId",
                    "label",
                    "summary",
                    "cycleIndex",
                    "isBest",
                    "scorecard",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "checkpointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/projects/{projectId}/agent-runs/{runId}/actions": {
      "get": {
        "operationId": "listAgentActions",
        "summary": "List agent actions.",
        "tags": [
          "agentRuns"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "runId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "cycleIndex": {
                        "type": "integer"
                      },
                      "actionType": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "rationale": {
                        "type": "string",
                        "nullable": true
                      },
                      "input": {
                        "nullable": true
                      },
                      "result": {
                        "nullable": true
                      },
                      "scoreDelta": {
                        "type": "number",
                        "nullable": true
                      },
                      "startedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "runId",
                      "cycleIndex",
                      "actionType",
                      "status",
                      "rationale",
                      "scoreDelta",
                      "startedAt",
                      "completedAt",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "title",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    }
  }
}